Bump Kafka dependency to 2.3 and replace NetworkClient inside MetadataClient with the public AdminClient interface
It is better to depend on public interfaces where possible to improve API compatibility with newer Kafka versions.
This patch refactors the MetadataClient, removing the use of the Kafka's internal NetworkClient class and the associated network.client.provider.class config.
It also changes the metadata.max.age.ms config from a Long to an Integer, as the AdminClient interface uses ints for its timeouts

+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/config/KafkaCruiseControlConfig.java
@@ -251,7 +251,7 @@ public class KafkaCruiseControlConfig extends AbstractConfig {
    */
   private void sanityCheckSamplingPeriod(Map<?, ?> originals) {
     long samplingIntervalMs = getLong(MonitorConfig.METRIC_SAMPLING_INTERVAL_MS_CONFIG);
-    long metadataTimeoutMs = getLong(MonitorConfig.METADATA_MAX_AGE_CONFIG);
+    long metadataTimeoutMs = getInt(MonitorConfig.METADATA_MAX_AGE_CONFIG);
     if (metadataTimeoutMs > samplingIntervalMs) {
       throw new ConfigException("Attempt to set metadata refresh timeout [" + metadataTimeoutMs +
                                 "] to be longer than sampling period [" + samplingIntervalMs + "].");
