+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/detector/AnomalyDetectorUtils.java
@@ -25,7 +25,7 @@ public class AnomalyDetectorUtils {
   private static final Logger LOG = LoggerFactory.getLogger(AnomalyDetectorUtils.class);
   public static final String KAFKA_CRUISE_CONTROL_OBJECT_CONFIG = "kafka.cruise.control.object";
   public static final String ANOMALY_DETECTION_TIME_MS_OBJECT_CONFIG = "anomaly.detection.time.ms.object";
-  public static final long MAX_METADATA_WAIT_MS = 60000L;
+  public static final int MAX_METADATA_WAIT_MS = 60000;
   public static final Anomaly SHUTDOWN_ANOMALY = new BrokerFailures();
 
   private AnomalyDetectorUtils() {
